blur: Make further optimizations
This removes the frame_pre_paint function that was adding padded damage.
As it turns out, this isn't needed at all. In addition, we don't actually
need to shrink the opaque region as it's simple not necessary. This means
even less resources are used when blurring surfaces with opaque region set.