No need for grid member variable anymore (#564)
I originally used it to keep track of views in the grid. Since find_view_in_grid() always goes through the list of views returned by get_views(), there is no need to keep this, we only need to save the number of rows, columns, and the number of views in the last row.